Make PageStart work Move more API to RazorPageBase
