+++ samples/MvcSandbox/Pages/Index.cshtml
@@ -1,5 +1,9 @@
 ﻿@page
 @model TestModel
+@{
+
+    ViewData["Title"] = "Hello from pages";
+}
 
 @functions {
