+++ samples/MvcSandbox/Pages/_PageStart.cshtml
@@ -1,0 +1,1 @@
+﻿@{ Layout = "_Layout";}
